+++ distributor-node/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 2.2.0
+
+- Updates `operator:set-metadata` CLI command to set distributor-node's operational status along with other metadata.
+- Adds `leader:set-node-operational-status` CLI command to set operational status of any distributor-node by Lead.
+
 ### 2.1.1
 
 - Bump deps @polkadot/api v10.7.1